move sendAddToChannelEphemeralPost to actions folder
