+++ app/components/post_add_channel_member/index.js
@@ -4,9 +4,6 @@
 import {connect} from 'react-redux';
 import {bindActionCreators} from 'redux';
 
-import {Posts} from 'mattermost-redux/constants';
-import {PostTypes} from 'mattermost-redux/action_types';
-
 import {addChannelMember} from 'mattermost-redux/actions/channels';
 import {removePost} from 'mattermost-redux/actions/posts';
 
@@ -14,7 +11,7 @@ import {getPost} from 'mattermost-redux/selectors/entities/posts';
 import {getChannel} from 'mattermost-redux/selectors/entities/channels';
 import {getCurrentUser} from 'mattermost-redux/selectors/entities/users';
 
-import {generateId} from 'app/utils/file';
+import {sendAddToChannelEphemeralPost} from 'app/actions/views/post';
 
 import PostAddChannelMember from './post_add_channel_member';
 
@@ -35,44 +32,12 @@ function mapStateToProps(state, ownProps) {
     };
 }
 
-function sendEphemeralPost(user, addedUsername, message, channelId, postRootId) {
-    return async (dispatch) => {
-        const timestamp = Date.now();
-        const post = {
-            id: generateId(),
-            user_id: user.id,
-            channel_id: channelId,
-            message,
-            type: Posts.POST_TYPES.EPHEMERAL_ADD_TO_CHANNEL,
-            create_at: timestamp,
-            update_at: timestamp,
-            root_id: postRootId,
-            parent_id: postRootId,
-            props: {
-                username: user.username,
-                addedUsername,
-            },
-        };
-
-        dispatch({
-            type: PostTypes.RECEIVED_POSTS,
-            data: {
-                order: [],
-                posts: {
-                    [post.id]: post,
-                },
-            },
-            channelId,
-        });
-    };
-}
-
 function mapDispatchToProps(dispatch) {
     return {
         actions: bindActionCreators({
             addChannelMember,
             removePost,
-            sendEphemeralPost,
+            sendAddToChannelEphemeralPost,
         }, dispatch),
     };
 }
